Add alacritty and remove base16
Add configurations for Alacritty terminal emulator.

Remove base16 and start using individual themes for Neovim and
Alacritty. Most of the good themes have a version for both, so we can
still use zsh to update themes on the fly and sync changes with Neovim.

Above changes mean refactoring multiple files while adding some new.
There are quite many dependencies, so this is a big one:
- remove all base16 dependencies
- add alacritty configuration and some themes
- add zsh function to change alacritty theme from cli
- sync neovim colorscheme with cli